Mark many class methods as const

+++ TKMovesets2/main.cpp
@@ -356,7 +356,7 @@ int WinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _In_  LP
 			glfwSwapBuffers(window);
 
 			if (io.DeltaTime < (1.0f / MAX_FPS)) {
-				Sleep(((1.0f / MAX_FPS) - io.DeltaTime) * 1000);
+				Sleep((DWORD)(((1.0f / MAX_FPS) - io.DeltaTime) * 1000));
 			}
 		}
